refactor(fs): eliminate srcShared/ namespace
Problem:
The `srcShared/` concept is not necessary. Instead, a more tractable and
less-churn approach is:
1. Assume that ALL modules are "cross platform" (i.e. they support both
   nodejs and web-mode).
2. Web-only logic lives in a `web/` directory relative to its namespace
   (or in some rare cases, a `fooWeb.ts` module).
2. Nodejs-only logic lives in a `nodejs/` directory relative to its namespace
   (or in some rare cases, a `fooNodejs.ts` module).

See also:
- https://github.com/aws/aws-toolkit-vscode/pull/5046#discussion_r1614541013
- https://github.com/aws/aws-toolkit-vscode/pull/5248#discussion_r1661084971

Solution:
- Move `fs.ts` into `shared/`.
- Delete `srcShared/`.
