+++ packages/core/src/codewhisperer/service/transformByQ/humanInTheLoopManager.ts
@@ -6,7 +6,7 @@ import * as vscode from 'vscode'
 import * as os from 'os'
 import path from 'path'
 import { FolderInfo, transformByQState } from '../../models/model'
-import { fsCommon } from '../../../srcShared/fs'
+import { fsCommon } from '../../../shared'
 import { createPomCopy, replacePomVersion } from './transformFileHandler'
 import { IManifestFile } from '../../../amazonqFeatureDev/models'
 import { getLogger } from '../../../shared/logger'
