+++ packages/core/src/codewhisperer/vue/backend.ts
@@ -11,7 +11,7 @@ import { VueWebview } from '../../webviews/main'
 import { isCloud9 } from '../../shared/extensionUtilities'
 import globals from '../../shared/extensionGlobals'
 import { telemetry, CodewhispererLanguage, CodewhispererGettingStartedTask } from '../../shared/telemetry/telemetry'
-import { fsCommon } from '../../srcShared/fs'
+import { fsCommon } from '../../shared'
 import { getLogger } from '../../shared/logger'
 import { AmazonQPromptSettings } from '../../shared/settings'
 import { CodeWhispererSource } from '../commands/types'
