+++ packages/core/src/shared/index.ts
@@ -42,3 +42,4 @@ export * from './utilities/pathUtils'
 export * from './errors'
 export * as messages from './utilities/messages'
 export * as errors from './errors'
+export { FileSystemCommon, fsCommon, fsCommon as fs } from './fs/fs'
