--- packages/core/src/srcShared/README.md
@@ -1,4 +1,0 @@
-# srcShared
-
-This folder must only contain code which can be used in the browser OR desktop.
-This is why we call it "shared".
