--- packages/core/src/srcShared/index.ts
@@ -1,6 +1,0 @@
-/*!
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export { FileSystemCommon, fsCommon, fsCommon as fs } from './fs'
