+++ packages/core/src/test/srcShared/fs.test.ts
@@ -9,7 +9,7 @@ import * as path from 'path'
 import * as utils from 'util'
 import { existsSync, mkdirSync, promises as nodefs, readFileSync, rmSync } from 'fs'
 import { FakeExtensionContext } from '../fakeExtensionContext'
-import fs from '../../srcShared/fs'
+import fs from '../../shared/fs/fs'
 import * as os from 'os'
 import { isMinVscode, isWin } from '../../shared/vscode/env'
 import Sinon from 'sinon'
